Fix spinner sync with upload operation termination. Made network service and gallery presenter the responders to finishing the operation, instead of cell presenter. Cell presenter still starts it, and load the cell's image.
This was not enough to fix the bug, as I was only querying the spinner when reusing the cell and in cell presenter, when reusing before operation finished, cell presenter is different, so no one is notified of the end. The above solved this for the reuse case but not for the not-reused but also not displayed, which fails cellForItem retrieval. That is why I added the willDisplay and didEndDisplaying cell methods in Gallery VC

This solved all bugs related to spinner sync. Printing will be deleted after tested again.